Cleanup SDK interface (#1479)
* Cleanup SDK interface

* Exclude protobuf/permissions

* Use uint16 as QuorumType

+++ shared_model/interfaces/common_objects/types.hpp
@@ -63,7 +63,7 @@ namespace shared_model {
       /// Permission set
       using PermissionSetType = std::set<PermissionNameType>;
       /// Type of Quorum used in transaction and set quorum
-      using QuorumType = uint32_t;
+      using QuorumType = uint16_t;
       /// Type of signature range, which returns when signatures are invoked
       using SignatureRangeType = boost::any_range<const interface::Signature &,
                                                   boost::forward_traversal_tag>;
